refactor: move item editor
