+++ app/src/features/items/index.ts
@@ -2,3 +2,4 @@ export * from './relations'
 export * from './details'
 export * from './table'
 export * from './useNavigateToItem'
+export * from './editor'
